Fix integer overflow in size computations across VP8/VP9 encoder
Harden 30+ integer overflow sites in pixel-dimension arithmetic,
buffer allocation size calculations, and PSNR metric computations.

The root cause is that cm->width * cm->height (both int) and similar
dimension-derived multiplications are performed in 32-bit signed
arithmetic throughout the codebase, despite VP9 supporting dimensions
up to 65536x65536. This creates undefined behavior (signed overflow),
incorrect allocation sizes, and a division-by-zero in rate control.

This is the same class of vulnerability that led to CVE-2023-5217
(heap buffer overflow in VP8 encoding, exploited as a 0-day) and
CVE-2024-5197 (integer overflow in vpx_img_alloc).

Changes:
- Add vpx_size_checked_mul() overflow-checked helpers (vpx_mem.h)
- Fix VP8 vp8_yv12_realloc_frame_buffer to use uint64_t like VP9's
- Fix get_token_alloc() int return type overflow (int -> int64_t)
- Fix cx_data_sz overflow in VP8 and VP9 encoder init
- Fix VP8 token allocation overflow in onyx_if.c
- Fix PSNR uint32_t samples overflow in vpx_dsp/psnr.c
- Fix 23 cm->width * cm->height signed overflow sites across 6 files

Bug: b/none

+++ vp9/encoder/vp9_encoder.h
@@ -1064,7 +1064,7 @@ static INLINE YV12_BUFFER_CONFIG *get_ref_frame_buffer(
                                 : NULL;
 }
 
-static INLINE int get_token_alloc(int mb_rows, int mb_cols) {
+static INLINE int64_t get_token_alloc(int mb_rows, int mb_cols) {
   // TODO(JBB): double check we can't exceed this token count if we have a
   // 32x32 transform crossing a boundary at a multiple of 16.
   // mb_rows, cols are in units of 16 pixels. We assume 3 planes all at full
@@ -1076,7 +1076,7 @@ static INLINE int get_token_alloc(int mb_rows, int mb_cols) {
       ALIGN_POWER_OF_TWO(mb_rows, MI_BLOCK_SIZE_LOG2 - 1);
   const int aligned_mb_cols =
       ALIGN_POWER_OF_TWO(mb_cols, MI_BLOCK_SIZE_LOG2 - 1);
-  return aligned_mb_rows * aligned_mb_cols * (16 * 16 * 3 + 4);
+  return (int64_t)aligned_mb_rows * aligned_mb_cols * (16 * 16 * 3 + 4);
 }
 
 // Get the allocated token size for a tile. It does the same calculation as in
@@ -1085,7 +1085,7 @@ static INLINE int allocated_tokens(TileInfo tile) {
   int tile_mb_rows = (tile.mi_row_end - tile.mi_row_start + 1) >> 1;
   int tile_mb_cols = (tile.mi_col_end - tile.mi_col_start + 1) >> 1;
 
-  return get_token_alloc(tile_mb_rows, tile_mb_cols);
+  return (int)get_token_alloc(tile_mb_rows, tile_mb_cols);
 }
 
 static INLINE void get_start_tok(VP9_COMP *cpi, int tile_row, int tile_col,
