+++ vpx_dsp/psnr.c
@@ -188,14 +188,14 @@ void vpx_calc_highbd_psnr(const YV12_BUFFER_CONFIG *a,
   const int b_strides[3] = { b->y_stride, b->uv_stride, b->uv_stride };
   int i;
   uint64_t total_sse = 0;
-  uint32_t total_samples = 0;
+  uint64_t total_samples = 0;
   const double peak = (double)((1 << in_bit_depth) - 1);
   const unsigned int input_shift = bit_depth - in_bit_depth;
 
   for (i = 0; i < 3; ++i) {
     const int w = widths[i];
     const int h = heights[i];
-    const uint32_t samples = w * h;
+    const uint64_t samples = (uint64_t)w * h;
     uint64_t sse;
     if (a->flags & YV12_FLAG_HIGHBITDEPTH) {
       if (input_shift) {
@@ -209,7 +209,7 @@ void vpx_calc_highbd_psnr(const YV12_BUFFER_CONFIG *a,
       sse = get_sse(a_planes[i], a_strides[i], b_planes[i], b_strides[i], w, h);
     }
     psnr->sse[1 + i] = sse;
-    psnr->samples[1 + i] = samples;
+    psnr->samples[1 + i] = (unsigned int)samples;
     psnr->psnr[1 + i] = vpx_sse_to_psnr(samples, peak, (double)sse);
 
     total_sse += sse;
@@ -217,7 +217,7 @@ void vpx_calc_highbd_psnr(const YV12_BUFFER_CONFIG *a,
   }
 
   psnr->sse[0] = total_sse;
-  psnr->samples[0] = total_samples;
+  psnr->samples[0] = (unsigned int)total_samples;
   psnr->psnr[0] =
       vpx_sse_to_psnr((double)total_samples, peak, (double)total_sse);
   psnr->spatial_layer_id = spatial_layer_id;
@@ -237,16 +237,16 @@ void vpx_calc_psnr(const YV12_BUFFER_CONFIG *a, const YV12_BUFFER_CONFIG *b,
   const int b_strides[3] = { b->y_stride, b->uv_stride, b->uv_stride };
   int i;
   uint64_t total_sse = 0;
-  uint32_t total_samples = 0;
+  uint64_t total_samples = 0;
 
   for (i = 0; i < 3; ++i) {
     const int w = widths[i];
     const int h = heights[i];
-    const uint32_t samples = w * h;
+    const uint64_t samples = (uint64_t)w * h;
     const uint64_t sse =
         get_sse(a_planes[i], a_strides[i], b_planes[i], b_strides[i], w, h);
     psnr->sse[1 + i] = sse;
-    psnr->samples[1 + i] = samples;
+    psnr->samples[1 + i] = (unsigned int)samples;
     psnr->psnr[1 + i] = vpx_sse_to_psnr(samples, peak, (double)sse);
 
     total_sse += sse;
@@ -254,7 +254,7 @@ void vpx_calc_psnr(const YV12_BUFFER_CONFIG *a, const YV12_BUFFER_CONFIG *b,
   }
 
   psnr->sse[0] = total_sse;
-  psnr->samples[0] = total_samples;
+  psnr->samples[0] = (unsigned int)total_samples;
   psnr->psnr[0] =
       vpx_sse_to_psnr((double)total_samples, peak, (double)total_sse);
   psnr->spatial_layer_id = spatial_layer_id;
